+++ cms/envs/common.py
@@ -581,6 +581,9 @@ MODULESTORE = {
         }
     }
 }
+# How many seconds should we wait before assuming a course publish has made it
+# to the MongoDB secondaries (and is therefore safe to read).
+MODULESTORE_PUBLISH_SIGNAL_DELAY = 30
 
 # Modulestore-level field override providers. These field override providers don't
 # require student context.
