+++ common/djangoapps/django_comment_common/tests.py
@@ -4,6 +4,7 @@ from opaque_keys.edx.locator import CourseLocator
 from six import text_type
 
 from django_comment_common.models import Role
+import django_comment_common.signals
 from models import CourseDiscussionSettings
 from openedx.core.djangoapps.course_groups.cohorts import CourseCohortsSettings
 from student.models import CourseEnrollment, User
@@ -11,7 +12,8 @@ from utils import get_course_discussion_settings, set_course_discussion_settings
 from xmodule.modulestore import ModuleStoreEnum
 from xmodule.modulestore.django import modulestore
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
-from xmodule.modulestore.tests.factories import CourseFactory
+from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
+import openedx.core.djangoapps.request_cache as request_cache
 
 
 @attr(shard=1)
@@ -135,3 +137,39 @@ class CourseDiscussionSettingsTest(ModuleStoreTestCase):
                 text_type(value_error.exception),
                 exception_msg_template.format(field['name'], field['type'].__name__)
             )
+
+
+class CoursePublishHandlerTestCase(ModuleStoreTestCase):
+    """
+    Tests for discussion updates on course publish.
+    """
+    ENABLED_SIGNALS = ['course_published']
+
+    def test_discussion_id_map_updates_on_publish(self):
+        course_key_args = dict(org='org', course='number', run='run')
+        course_key = self.store.make_course_key(**course_key_args)
+
+        with self.assertRaises(CourseDiscussionSettings.DoesNotExist):
+            CourseDiscussionSettings.objects.get(course_id=course_key)
+
+        # create course
+        course = CourseFactory(emit_signals=True, **course_key_args)
+        self.assertEqual(course.id, course_key)
+        self._assert_discussion_id_map(course_key, {})
+
+        # create discussion block
+        request_cache.clear_cache(name=None)
+        discussion_id = 'discussion1'
+        discussion_block = ItemFactory.create(
+            parent_location=course.location,
+            category="discussion",
+            discussion_id=discussion_id,
+        )
+        self._assert_discussion_id_map(course_key, {discussion_id: str(discussion_block.location)})
+
+    def _assert_discussion_id_map(self, course_key, expected_map):
+        """
+        Verifies the discussion ID map for the given course matches the expected value.
+        """
+        discussion_settings = CourseDiscussionSettings.objects.get(course_id=course_key)
+        self.assertDictEqual(discussion_settings.discussions_id_map, expected_map)
